% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yg_colors.R
\name{yg_palette}
\alias{yg_palette}
\title{YouGov color palette generator}
\usage{
yg_palette("name", n, "type")
}
\arguments{
\item{name}{Name of the specific palette in quotation marks.}

\item{n}{Number of colors to select from the palette. If null, then all colors in the palette are selected.}

\item{type}{Specify the type of color mapping, either "continuous" or "discrete" in quotation marks. Use "continuous" to include more colors than those in the palette. See \code{examples} below for more.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colors
}
\description{
YouGov color palette generator
}
\examples{
# Display each palette
yg_palette("ps_core")
yg_palette("secondary_core")
yg_palette("grapefruit")
yg_palette("plum")
yg_palette("pomegranate")
yg_palette("blueberry")
yg_palette("avocado")
yg_palette("full")
yg_palette("grayscale")

# Interpolating between existing colors based on the palettes using the "continuous" type
yg_palette(50, name = "ps_core", type = "continuous")
yg_palette(50, name = "secondary_core", type = "continuous")
yg_palette(50, name = "grapefruit", type = "continuous")
yg_palette(50, name = "plum", type = "continuous")
yg_palette(50, name = "pomegranate", type = "continuous")
yg_palette(50, name = "blueberry", type = "continuous")
yg_palette(50, name = "avocado", type = "continuous")
yg_palette(50, name = "full", type = "continuous")
yg_palette(50, name = "grayscale", type = "continuous")
}
\references{
Philip Waggoner. 2019. amerika: An American politics-inspired color palette generator. R package version 0.1.

Karthik Ram and Hadley Wickham. 2015. wesanderson: a Wes Anderson palette generator. R package version 0.3.
}
