% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yg_colors.R
\docType{data}
\name{yg_palettes}
\alias{yg_palettes}
\title{YouGov color palettes}
\format{
An object of class \code{list} of length 9.
}
\usage{
yg_palettes
}
\description{
Use \code{\link{yg_palette}} to produce desired color palette
}
\keyword{datasets}
